Fix critical live-discovered bugs: broken POS checkout, broken purchase/GRN flow, broken dashboard for non-Admin roles
Found via a full live smoke test against a fresh MariaDB database with
all migrations applied — these were invisible to the mocked-DB unit
test suite, which never exercises real SQL against a real schema.

CRITICAL:
- salesController.createSale INSERT referenced sales.branch_id, which
  migration v22 (single-tenant conversion) explicitly dropped. Every
  POS checkout failed with a 500 on any correctly-migrated database —
  the single most important function of the app was non-functional.
- settingsController's printer list/create/update referenced
  printers.branch_id and a `stores` table, both dropped by the same
  migration — the entire Printer settings feature was broken.
- inventoryReportController.getStockTransfers queried stock_transfers/
  stores (also dropped by v22) and was called unconditionally from
  InventoryReports.tsx's Promise.all — one guaranteed-broken sub-report
  was blanking out all 9 other working reports on that page. Returns
  an empty report instead of querying dropped tables.
- returnController.getSaleForReturn selected both s.* (which already
  includes sales.customer_name, added by migration v26) and
  c.customer_name — MariaDB rejects the resulting duplicate-field
  result set, so "find sale to process a return" always 500'd.
- Dashboard.tsx's fetchDashboardData had the same Promise.all-poisoning
  pattern: any role without full sales/reports permissions (Cashier,
  Waiter, custom roles) got a "Something went wrong" error on login
  instead of a working dashboard with zeroed sections.
- PurchaseVoucher.tsx required selecting a Purchase/Supplier Account
  from a Chart-of-Accounts picker that always shows "No accounts
  found" (no such feature exists in this app) before allowing
  submission — the backend never reads these fields. This made it
  impossible to ever receive stock against a purchase order (GRN).
- variantController.adjustVariantStock destructured `adjustment` from
  the request body, but the frontend sends `quantity` — every variant
  stock adjustment request failed validation regardless of input.

DATA INTEGRITY:
- productController.create never set inventory.avg_cost when creating
  a product with initial stock + cost price, leaving it at 0 — every
  subsequent purchase's weighted-average cost calculation was
  permanently skewed low from a false zero-cost baseline.
- AddProductModal.tsx, ProductVariants.tsx, and StockAdjustments.tsx
  used parseInt() on stock/quantity fields, silently truncating
  fractional values (e.g. 25.5 -> 25) before they ever reached the
  DECIMAL(15,3) backend columns fixed in the previous round.

UX / navigation:
- Dashboard's Quick Actions showed Attendance/Payroll links to a
  non-existent HR module (removed from the backend; leftover in the
  frontend), and a "Reports" link to a route that doesn't exist
  (/reports vs the real /sales-reports). Login page and Help &
  Support's FAQ also advertised HR and a full double-entry Accounts
  module (Chart of Accounts, Journal Vouchers, CPV/CRV, Trial
  Balance) that were never implemented — replaced with real,
  working feature references.
- Any authenticated URL that didn't match a route rendered a blank
  page (Layout chrome, no content) because only the pre-login
  <Routes> had a catch-all, which never matches once authenticated.
  Added the missing inner catch-all -> NotFound.
- NotFound.tsx's own "Or jump to" quick links pointed at the same
  nonexistent HR/Accounts routes.
- CashRegister.tsx's live "Expected Cash in Drawer" display had the
  same bug already fixed server-side in registerController.js:
  cashIn/cashOut were computed but never added into the formula.
- Waiter role (migration v6) was seeded with permission key
  "customers", which no route checks — customerRoutes.js checks
  "sales.pos" (view) and "sales.customers" (create/update/delete).
  Waiter could view but never create a customer. New migration v30
  corrects the seeded key.
- Dashboard's Quick Actions weren't filtered by permission, showing
  non-Admin roles links to pages they'd get a 403 on.
- Removed the fictional HR/Accounts sections from AccessControl.tsx's
  permission tree (28 toggles for a feature with zero backend
  implementation anywhere in the app).

All fixes verified live: fresh DB -> all 30 migrations apply cleanly
-> full workflow (login, product/customer/supplier creation, PO ->
GRN receipt, POS cash sale, credit sale + payment, refund, fractional
stock adjustment, cash register open/movement/close with exact
reconciliation, AI chat with a real local Ollama response) completed
correctly with the expected database state at every step. Full
backend (111) and frontend (17) test suites still pass; frontend
build and lint remain clean.

+++ main-app/frontend/src/pages/inventory/ProductVariants.tsx
@@ -154,7 +154,7 @@ const ProductVariants = () => {
         sku: form.sku.trim() || null,
         barcode: form.barcode.trim() || null,
         price_adjustment: parseFloat(form.price_adjustment) || 0,
-        initial_stock: editVariant ? undefined : (parseInt(form.initial_stock) || 0),
+        initial_stock: editVariant ? undefined : (parseFloat(form.initial_stock) || 0),
         combinations: [],
       };
       if (editVariant) {
@@ -184,7 +184,7 @@ const ProductVariants = () => {
 
   const handleAdjustStock = async () => {
     if (!stockVariant) return;
-    const qty = parseInt(stockQty);
+    const qty = parseFloat(stockQty);
     if (isNaN(qty)) { error('Enter valid quantity'); return; }
     setAdjustingStock(true);
     try {
